React Router Multi Layout 적용
- 멀티 레이아웃 적용을 위한 전체적인 리팩토링 진행.
- 두 개의 레이아웃
  1. 기본 레이아웃
  2. 관리자 대시보드 레이아웃
- 디렉토리 구조 리팩토링.
```
./src
├── components: UI 컴포넌트 모음.
├── containers: 데이터를 컨트롤하는 컨테이너 모음.
├── services: 외부 API 및 DOM API 관련 유틸 모음.
└── router: 서비스 라우팅 정보.
```
