+++ src/containers/commons/Nav/MainMenus.tsx
@@ -5,16 +5,18 @@ import List from '@mui/material/List';
 import AccountCircleOutlinedIcon from '@mui/icons-material/AccountCircleOutlined';
 import BlockOutlinedIcon from '@mui/icons-material/BlockOutlined';
 
-import LabelMenu from './LabelMenu';
+import LinkedMenu from './LinkedMenu';
 
 const menuList = [
   {
     key: 'user',
+    to: '/user',
     label: '사용자',
     icon: <AccountCircleOutlinedIcon />,
   },
   {
     key: 'dummy',
+    to: '/dummy',
     label: '더미메뉴',
     icon: <BlockOutlinedIcon />,
   },
@@ -23,9 +25,12 @@ const menuList = [
 function MainMenus() {
   return (
     <List>
-      {menuList.map(({ key, label, icon }) => (
-        <LabelMenu
+      {menuList.map(({
+        key, to, label, icon,
+      }) => (
+        <LinkedMenu
           key={key}
+          href={to}
           label={label}
           icon={icon}
         />
