+++ src/containers/layouts/Main.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default ({ children }: any) => (
+  <div>
+    {children}
+  </div>
+);
