+++ src/containers/views/Admin/User/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Box from '@mui/material/Box';
+
+import UserContainer from './container/UserContainer';
+
+export default () => (
+  <Box
+    sx={{ display: 'flex' }}
+  >
+    <UserContainer />
+  </Box>
+);
