+++ src/containers/views/Admin/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Box from '@mui/material/Box';
+
+export default () => (
+  <Box
+    sx={{ display: 'flex' }}
+  >
+    관리자 페이지 입니다.
+  </Box>
+);
