--- src/domains/administrator/domains/user/UserPage.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import UserContainer from './container/UserContainer';
-
-function UserPage() {
-  return (
-    <UserContainer />
-  );
-}
-
-export default UserPage;
