+++ src/index.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
+
 import ReactDOM from 'react-dom';
+
 import {
   BrowserRouter,
 } from 'react-router-dom';
